// src/types/SessionMeta.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SessionActivity } from "./SessionActivity";

export type SessionMeta = { id: string, projectId: number, title: string, messageCount: number, lastModified: number, gitBranch: string | null, cwd: string | null, activity: SessionActivity, };